+++ dflib/src/main/java/org/dflib/exp/AsExp.java
@@ -9,7 +9,7 @@ import java.util.Objects;
 public class AsExp<T> implements Exp<T>  {
 
     private final Exp<T> delegate;
-    private final String name;
+    protected final String name;
 
     public AsExp(String name, Exp<T> delegate) {
         this.name = name;