+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #359 Exp: IllegalArgumentException on arithmetics ops following aggregation
 * #361 Introduce aggregating semantics in Exp
+* #372 NumExp.as() should return a NumExp
 
 ## Release 1.1.0